services/korean-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
python-dotenv==1.0.0
google-generativeai==0.8.5
google-cloud-vision==3.4.5
celery==5.3.4
redis==5.0.1
flower==2.0.1
requests==2.31.0
httpx==0.25.0
Pillow==10.1.0
pytesseract==0.3.10
opencv-python==4.8.1.78
numpy==1.24.3
pandas==2.1.4